refactor page codes to match new structure
